Refactor: [prototype] remove freezing of prototypes
Freezing prototype objects on extending could turn out to be a performance
bottleneck in the long run. Also it is redundant as we need to freeze each
extension recursively.

Implementing the immutability of contained data and container prototype will be
a responsible of the final level containers thatare exposed to the public.